requirements.txt
boto3==1.10.7
botocore==1.13.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.17
numpy==1.17.3
protobuf==3.6.0
pycparser==2.19
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.35
sentencepiece==0.1.83
six==1.12.0
torch==1.3.0
transformers==2.1.1
urllib3==1.25.6
Werkzeug==0.16.0
